kernel-engine/requirements.txt
# ── Core Web Framework ────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.13.0

# ── Validation & Config ───────────────────────────────────────────────────────
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.0

# ── HTTP Client ───────────────────────────────────────────────────────────────
httpx[http2]>=0.27.0

# ── Feature 1: Async Task Queue ───────────────────────────────────────────────
celery[redis]>=5.4.0
redis>=5.0.0

# ── Feature 3: System Load Monitoring & Cloud LLM Routing ────────────────────
psutil>=5.9.0
google-generativeai>=0.8.0
anthropic>=0.40.0

# ── Feature 5: Web Fetcher ────────────────────────────────────────────────────
beautifulsoup4>=4.12.0
lxml>=5.0.0